feat: loads submenu
