+++ src/lib/Ggraphic/index.svelte
@@ -211,6 +211,7 @@
   .graphic-container {
     width: 100%;
     height: 100%;
+    background-color: white;
   }
 
   .line {
