+++ src/lib/Options/LoadCases/LoadCases.svelte
@@ -1,17 +1,22 @@
 <script>
-  import List, { Item, Separator, Text } from "@smui/list";
-  import { loads } from "../../store";
-  import { _ } from "svelte-i18n";
+  import Button, { Label, Icon } from "@smui/button";
+  import List, { Item, Separator } from "@smui/list";
   import LoadCase from "./LoadCase.svelte";
   import LoadCaseForm from "./LoadCaseForm.svelte";
+  import { _ } from "svelte-i18n";
+  import { loads } from "$lib/store";
+  import { menuRoute } from "../menuRouter";
 </script>
+
+<Button on:click={() => menuRoute.set("main")}>
+  <Icon class="material-icons">arrow_back</Icon>
+  <Label>Back</Label>
+</Button>
 
 <div class="loads">
   <List>
     {#each $loads as _, i}
-      <Item
-        style="padding: 10px; text-align: center; justify-content: center"
-      >
+      <Item style="padding: 10px; text-align: center; justify-content: center">
         <LoadCase loadId={i} />
       </Item>
       <Separator />
