+++ src/lib/Options/menuRouter.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const menuRoute = writable<"main" | "loads">("main");
